GridAnalysis

Documentation for GridAnalysis.

GridAnalysis.DayAhead

GridAnalysis.MarketSimulator

GridAnalysis.UCED

GridAnalysis.plot_demand_stack

GridAnalysis.plot_generation_stack

GridAnalysis.plot_net_demand_stack

GridAnalysis.plot_net_demand_stack_prev

GridAnalysis.plot_prices

GridAnalysis.plot_thermal_commit

PowerSimulations.Simulation

GridAnalysis.MEC

GridAnalysis.bus_mapping

GridAnalysis.duals_constraint_names

GridAnalysis.evaluate_prices

GridAnalysis.evaluate_prices

GridAnalysis.evaluate_prices

GridAnalysis.evaluate_prices

GridAnalysis.fuel_type_mapping

GridAnalysis.get_bus_name

GridAnalysis.get_time_series_params

GridAnalysis.plot_demand_stack!

GridAnalysis.plot_demand_stack!

GridAnalysis.plot_generation_stack!

GridAnalysis.plot_generation_stack!

GridAnalysis.plot_net_demand_stack!

GridAnalysis.plot_net_demand_stack!

GridAnalysis.plot_net_demand_stack_prev!

GridAnalysis.plot_net_demand_stack_prev!

GridAnalysis.plot_prices!

GridAnalysis.plot_prices!

GridAnalysis.plot_thermal_commit!

GridAnalysis.plot_thermal_commit!

GridAnalysis.run_multiday_simulation

GridAnalysis.DayAhead — Type

DayAhead <: MarketSimulator

A type that represents DayAhead (DA) market clearings.

GridAnalysis.MarketSimulator — Type

MarketSimulator

A type to represent different market clearing structures.

GridAnalysis.UCED — Type

UCED <: DayAhead

Structure with two DA systems stored (system_uc and system_ed). One represents an Unit Commitment (UC) used to fix binary variables, the other an Economic Dispatch (ED) defined by template template_ed. Each model needs its own solver (solver_uc and solver_ed). system_ed and the resulting problem are defined as a heuristic to calculate DA energy prices.

GridAnalysis.plot_demand_stack — Method

plot_demand_stack(
    system::System,
    results::SimulationProblemResults)

Function to plot the Demand over the time period covered by the results. The bus_names controls which buses we want to include in the plot.

GridAnalysis.plot_generation_stack — Method

plot_generation_stack(
    system::System,
    results::SimulationProblemResults;
    generator_fields::AbstractArray=[:P__ThermalStandard, :P__RenewableDispatch],
    bus_names::AbstractArray=[])

Plot the generation mix over the time period covered by the results. The bus_names and generator_fields control which buses, and generator types we want to include the plot.

GridAnalysis.plot_net_demand_stack — Method

plot_net_demand_stack(
    system::System,
    results::SimulationProblemResults)

Function to plot the Demand over the time period covered by the results. The bus_names controls which buses we want to include in the plot. Renewable Dispatch data is the time series from the system.

GridAnalysis.plot_net_demand_stack_prev — Method

plot_net_demand_stack_prev(
    system::System,
    results::SimulationProblemResults)

Function to plot the Demand over the time period covered by the results. The bus_names controls which buses we want to include in the plot. It uses the Renewable Dispatch from the results.

GridAnalysis.plot_prices — Method

plot_prices(
    market_simulator::MarketSimulator,
    results::SimulationProblemResults;
    bus_names::AbstractArray=[])

Plot the simulation prices over the time period covered by the results. The bus_names control which buses we want to include the plot.

GridAnalysis.plot_thermal_commit — Method

plot_thermal_commit(
    system::System,
    results::SimulationProblemResults)

Function to plot the Thermal Standard Commit variables over the time period covered by the results. The results should be from the unit commitment problem. 1 is ON, 0 is OFF.

PowerSimulations.Simulation — Type

Simulation(simulator::UCED, initial_time::Date, steps::Int) -> PSI.Simulation

Builds a multiday UC-ED PSI.Simulation from a UCED market simulator, an initial date and number of simulation steps.

GridAnalysis.MEC — Method

MEC(system::System, problem_results::PSI.SimulationProblemResults)

Returns the Marginal Energy Component (MEC) of OPF problem's solution. This component is defined as the dual value (or lagrangian multiplier) of a grid-wide energy balance constraint (named in PSI CopperPlateBalance)

GridAnalysis.bus_mapping — Method

bus_mapping(system::System)

Returns generator to bus mapping.

GridAnalysis.duals_constraint_names — Method

duals_constraint_names(<:AbstractPowerModel)

Return the constraints for which we care about the duals (because they form the energy prices) when using a specified network formulation.

GridAnalysis.evaluate_prices — Method

evaluate_prices(::Type{CopperPlatePowerModel}, system::System, problem_results::PSI.SimulationProblemResults, kwargs)

Returns a grid-unique series of energy prices for the simulation's data-range. In this formulation, for each period of the problem, the energy prices are constructed using the dual value (or lagrangian multiplier) of a grid-wide energy balance constraint (named in PSI CopperPlateBalance).

GridAnalysis.evaluate_prices — Method

evaluate_prices(transmission::Type{StandardPTDFModel}, system::System, problem_results::PSI.SimulationProblemResults, kwargs)

Returns a nodal-wide series of energy prices (locational marginal prices - LMPS) for the simulation's data-range. In this formulation, for each period of the problem, the energy prices are constructed using the dual value (or lagrangian multiplier) of a grid-wide energy balance constraint (named in PSI CopperPlateBalance) and the dual values of a branch-wide power-flow constraint (named in PSI network_flow__Line).

GridAnalysis.evaluate_prices — Method

evaluate_prices(market_simulator::UCED, problem_results::PSI.SimulationProblemResults)

Returns energy prices for the simulation's data-range.

GridAnalysis.evaluate_prices — Method

evaluate_prices(::Union{Type{NFAPowerModel},Type{DCPPowerModel}}, system::System, problem_results::PSI.SimulationProblemResults, kwargs)

Returns a nodal-wide series of energy prices (locational marginal prices - LMPS) for the simulation's data-range. In this formulation, for each period of the problem, the energy prices are constructed using the dual values (or lagrangian multipliers) of each nodal energy balance constraint (named in PSI nodal_balance_active__Bus).

GridAnalysis.fuel_type_mapping — Method

fuel_type_mapping(system)

return a mapping between the bus name and the fuel type for the given `system`.

GridAnalysis.get_bus_name — Method

get_bus_name(gen::Generator)

Returns the bus name of a generator or load.

GridAnalysis.get_time_series_params — Method

get_time_series_params(system::System)

Returns the parameters associated with the time-series attached to the system.

GridAnalysis.plot_demand_stack! — Method

plot_demand_stack(
    system::System,
    results::SimulationProblemResults)

Function to plot the Demand over the time period covered by the results. The bus_names controls which buses we want to include in the plot.

GridAnalysis.plot_demand_stack! — Method

plot_demand_stack(
    system::System,
    results::SimulationProblemResults)

Function to plot the Demand over the time period covered by the results. The bus_names controls which buses we want to include in the plot.

GridAnalysis.plot_generation_stack! — Method

plot_generation_stack(
    system::System,
    results::SimulationProblemResults;
    generator_fields::AbstractArray=[:P__ThermalStandard, :P__RenewableDispatch],
    bus_names::AbstractArray=[])

Plot the generation mix over the time period covered by the results. The bus_names and generator_fields control which buses, and generator types we want to include the plot.

GridAnalysis.plot_generation_stack! — Method

plot_generation_stack(
    system::System,
    results::SimulationProblemResults;
    generator_fields::AbstractArray=[:P__ThermalStandard, :P__RenewableDispatch],
    bus_names::AbstractArray=[])

Plot the generation mix over the time period covered by the results. The bus_names and generator_fields control which buses, and generator types we want to include the plot.

GridAnalysis.plot_net_demand_stack! — Method

plot_net_demand_stack(
    system::System,
    results::SimulationProblemResults)

Function to plot the Demand over the time period covered by the results. The bus_names controls which buses we want to include in the plot. Renewable Dispatch data is the time series from the system.

GridAnalysis.plot_net_demand_stack! — Method

plot_net_demand_stack(
    system::System,
    results::SimulationProblemResults)

Function to plot the Demand over the time period covered by the results. The bus_names controls which buses we want to include in the plot. Renewable Dispatch data is the time series from the system.

GridAnalysis.plot_net_demand_stack_prev! — Method

plot_net_demand_stack_prev(
    system::System,
    results::SimulationProblemResults)

Function to plot the Demand over the time period covered by the results. The bus_names controls which buses we want to include in the plot. It uses the Renewable Dispatch from the results.

GridAnalysis.plot_net_demand_stack_prev! — Method

plot_net_demand_stack_prev(
    system::System,
    results::SimulationProblemResults)

Function to plot the Demand over the time period covered by the results. The bus_names controls which buses we want to include in the plot. It uses the Renewable Dispatch from the results.

GridAnalysis.plot_prices! — Method

plot_prices(
    market_simulator::MarketSimulator,
    results::SimulationProblemResults;
    bus_names::AbstractArray=[])

Plot the simulation prices over the time period covered by the results. The bus_names control which buses we want to include the plot.

GridAnalysis.plot_prices! — Method

plot_prices(
    market_simulator::MarketSimulator,
    results::SimulationProblemResults;
    bus_names::AbstractArray=[])

Plot the simulation prices over the time period covered by the results. The bus_names control which buses we want to include the plot.

GridAnalysis.plot_thermal_commit! — Method

plot_thermal_commit(
    system::System,
    results::SimulationProblemResults)

Function to plot the Thermal Standard Commit variables over the time period covered by the results. The results should be from the unit commitment problem. 1 is ON, 0 is OFF.

GridAnalysis.plot_thermal_commit! — Method

plot_thermal_commit(
    system::System,
    results::SimulationProblemResults)

Function to plot the Thermal Standard Commit variables over the time period covered by the results. The results should be from the unit commitment problem. 1 is ON, 0 is OFF.

GridAnalysis.run_multiday_simulation — Function

run_multiday_simulation(simulator::UCED, initial_time::Date, steps::Int) -> SimulationResults

Runs a multiday PSI.Simulation from a MarketSimulator, an initial date and number of simulation steps.