GridAnalysis

Documentation for GridAnalysis.

GridAnalysis.DayAhead

GridAnalysis.MarketSimulator

GridAnalysis.UCED

PowerSimulations.Simulation

GridAnalysis.run_multiday_simulation

GridAnalysis.DayAhead — Type

DayAhead <: MarketSimulator

A type that represents DayAhead (DA) market clearings.

GridAnalysis.MarketSimulator — Type

MarketSimulator

A type to represent different market clearing structures.

GridAnalysis.UCED — Type

UCED <: DayAhead

Structure with two DA systems stored (system_uc and system_ed). One represents an Unit Commitment (UC) used to fix binary variables, the other an Economic Dispatch (ED) defined by template template_ed. Each model needs its own solver (solver_uc and solver_ed). system_ed and the resulting problem are defined as a heuristic to calculate DA energy prices.

PowerSimulations.Simulation — Type

Simulation(simulator::UCED, initial_time::Date, steps::Int) -> PSI.Simulation

Builds a multiday UC-ED PSI.Simulation from a UCED market simulator, an initial date and number of simulation steps.

GridAnalysis.run_multiday_simulation — Function

run_multiday_simulation(simulator::UCED, initial_time::Date, steps::Int) -> SimulationResults

Runs a multiday PSI.Simulation from a MarketSimulator, an initial date and number of simulation steps.